+++ libcxx/test/benchmarks/containers/map.bench.cpp
@@ -14,7 +14,7 @@
 #include <random>
 #include <vector>
 
-#include "CartesianBenchmarks.h"
+#include "../CartesianBenchmarks.h"
 #include "benchmark/benchmark.h"
 #include "test_macros.h"