+++ lldb/unittests/Host/PipeTest.cpp
@@ -56,6 +56,8 @@ TEST_F(PipeTest, OpenAsReader) {
 #endif
 
 #if !defined(_AIX)
+// This test is flaky on Windows on Arm.
+#ifndef _WIN32
 TEST_F(PipeTest, WriteWithTimeout) {
   Pipe pipe;
   ASSERT_THAT_ERROR(pipe.CreateNew(false).ToError(), llvm::Succeeded());
@@ -152,3 +154,4 @@ TEST_F(PipeTest, WriteWithTimeout) {
                     llvm::Succeeded());
 }
 #endif
+#endif /*ifndef _WIN32*/